cms-api: Split DamModule into DamFilesModule, DamImagesModule and DamBlocksModule
DamModule registered file handling, image scaling, the block transformers and the
dependents resolver as one unit, so a project that only needs DAM file upload and
storage had to bring ImgproxyModule and DependenciesModule along with it.

Move the registrations into four sub-modules and turn DamModule into a facade that
composes all of them, so DamModule.register() behaves as before. FilesService
receives the dominant color calculator through the optional
DAM_DOMINANT_COLOR_CALCULATOR token, which DamImagesModule provides. Without that
module, uploads skip the color.
