+++ packages/api/cms-api/src/dam/dam.constants.ts
@@ -1,2 +1,3 @@
 export const DAM_CONFIG = "dam-config";
 export const DAM_FILE_VALIDATION_SERVICE = "dam-file-validation-service";
+export const DAM_DOMINANT_COLOR_CALCULATOR = "dam-dominant-color-calculator";
