+++ packages/api/cms-api/src/index.ts
@@ -183,13 +183,16 @@ export { damDefaultAcceptedMimetypes } from "./dam/common/mimeTypes/dam-default-
 export { DamConfig } from "./dam/dam.config";
 export { DAM_CONFIG } from "./dam/dam.constants";
 export { DamModule } from "./dam/dam.module";
+export { DamBlocksModule } from "./dam/dam-blocks.module";
+export { DamDependentsModule } from "./dam/dam-dependents.module";
+export { DamFilesModule } from "./dam/dam-files.module";
+export { DamImagesModule } from "./dam/dam-images.module";
 export { CreateFileInput, ImageFileInput, UpdateFileInput } from "./dam/files/dto/file.input";
 export { CreateFolderInput, UpdateFolderInput } from "./dam/files/dto/folder.input";
 export { DamFileAiContentType } from "./dam/files/entities/ai-content-type.enum";
 export { createFileEntity, FileInterface } from "./dam/files/entities/file.entity";
 export { DamFileImage } from "./dam/files/entities/file-image.entity";
 export { createFolderEntity, FolderInterface } from "./dam/files/entities/folder.entity";
-export { FileImagesResolver } from "./dam/files/file-image.resolver";
 export { FilesService } from "./dam/files/files.service";
 export { FoldersService } from "./dam/files/folders.service";
 export { DamDominantColorService } from "./dam/images/dam-dominant-color.service";
@@ -197,6 +200,7 @@ export { ImageInterface } from "./dam/images/dto/image.interface";
 export { HashImageParams, ImageParams } from "./dam/images/dto/image.params";
 export { ImageCropAreaInput } from "./dam/images/dto/image-crop-area.input";
 export { ImageCropArea } from "./dam/images/entities/image-crop-area.entity";
+export { FileImagesResolver } from "./dam/images/file-image.resolver";
 export { ImagesService } from "./dam/images/images.service";
 export { calculateInheritAspectRatio } from "./dam/images/images.util";
 export { IsAllowedImageAspectRatio, IsAllowedImageAspectRatioConstraint } from "./dam/images/validators/is-allowed-aspect-ratio.validator";
